Migrate more packs endpoints, refactor datastore tests
