+++ server/service/transport_packs.go
@@ -6,16 +6,6 @@ import (
 	"net/http"
 )
 
-func decodeDeletePackRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	name, err := stringFromRequest(r, "name")
-	if err != nil {
-		return nil, err
-	}
-	var req deletePackRequest
-	req.Name = name
-	return req, nil
-}
-
 func decodeDeletePackByIDRequest(ctx context.Context, r *http.Request) (interface{}, error) {
 	id, err := uintFromRequest(r, "id")
 	if err != nil {
@@ -34,14 +24,6 @@ func decodeGetPackRequest(ctx context.Context, r *http.Request) (interface{}, er
 	var req getPackRequest
 	req.ID = uint(id)
 	return req, nil
-}
-
-func decodeListPacksRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	opt, err := listOptionsFromRequest(r)
-	if err != nil {
-		return nil, err
-	}
-	return listPacksRequest{ListOptions: opt}, nil
 }
 
 func decodeApplyPackSpecsRequest(ctx context.Context, r *http.Request) (interface{}, error) {
